untracked API to Core
