+++ src/specklepy/api/credentials.py
@@ -9,37 +9,11 @@ from specklepy.logging import metrics
 from specklepy.logging.exceptions import SpeckleException
 from specklepy.transports.sqlite import SQLiteTransport
 
-
-class UserInfo(BaseModel):
-    name: Optional[str] = None
-    email: Optional[str] = None
-    company: Optional[str] = None
-    id: Optional[str] = None
-
-
-class Account(BaseModel):
-    isDefault: bool = False
-    token: Optional[str] = None
-    refreshToken: Optional[str] = None
-    serverInfo: ServerInfo = Field(default_factory=ServerInfo)
-    userInfo: UserInfo = Field(default_factory=UserInfo)
-    id: Optional[str] = None
-
-    def __repr__(self) -> str:
-        return (
-            f"Account(email: {self.userInfo.email}, server: {self.serverInfo.url},"
-            f" isDefault: {self.isDefault})"
-        )
-
-    def __str__(self) -> str:
-        return self.__repr__()
-
-    @classmethod
-    def from_token(cls, token: str, server_url: str = None):
-        acct = cls(token=token)
-        acct.serverInfo.url = server_url
-        return acct
-
+# following imports seem to be unnecessary, but they need to stay 
+# to not break the scripts using these functions as non-core
+from specklepy.core.api.credentials import (UserInfo, Account, StreamWrapper, 
+                                            get_account_from_token,
+                                            get_local_accounts as core_get_local_accounts)
 
 def get_local_accounts(base_path: Optional[str] = None) -> List[Account]:
     """Gets all the accounts present in this environment
@@ -51,41 +25,7 @@ def get_local_accounts(base_path: Optional[str] = None) -> List[Account]:
         List[Account] -- list of all local accounts or an empty list if
         no accounts were found
     """
-    accounts: List[Account] = []
-    try:
-        account_storage = SQLiteTransport(scope="Accounts", base_path=base_path)
-        res = account_storage.get_all_objects()
-        account_storage.close()
-        if res:
-            accounts.extend(Account.parse_raw(r[1]) for r in res)
-    except SpeckleException:
-        # cannot open SQLiteTransport, probably because of the lack
-        # of disk write permissions
-        pass
-
-    json_acct_files = []
-    json_path = str(speckle_path_provider.accounts_folder_path())
-    try:
-        os.makedirs(json_path, exist_ok=True)
-        json_acct_files.extend(
-            file for file in os.listdir(json_path) if file.endswith(".json")
-        )
-
-    except Exception:
-        # cannot find or get the json account paths
-        pass
-
-    if json_acct_files:
-        try:
-            accounts.extend(
-                Account.parse_file(os.path.join(json_path, json_file))
-                for json_file in json_acct_files
-            )
-        except Exception as ex:
-            raise SpeckleException(
-                "Invalid json accounts could not be read. Please fix or remove them.",
-                ex,
-            ) from ex
+    accounts = core_get_local_accounts(base_path)
 
     metrics.track(
         metrics.ACCOUNTS,
@@ -97,7 +37,6 @@ def get_local_accounts(base_path: Optional[str] = None) -> List[Account]:
 
     return accounts
 
-
 def get_default_account(base_path: Optional[str] = None) -> Optional[Account]:
     """
     Gets this environment's default account if any. If there is no default,
@@ -108,7 +47,7 @@ def get_default_account(base_path: Optional[str] = None) -> Optional[Account]:
     Returns:
         Account -- the default account or None if no local accounts were found
     """
-    accounts = get_local_accounts(base_path=base_path)
+    accounts = core_get_local_accounts(base_path=base_path)
     if not accounts:
         return None
 
@@ -119,42 +58,4 @@ def get_default_account(base_path: Optional[str] = None) -> Optional[Account]:
     metrics.initialise_tracker(default)
 
     return default
-
-
-def get_account_from_token(token: str, server_url: str = None) -> Account:
-    """Gets the local account for the token if it exists
-    Arguments:
-        token {str} -- the api token
-
-    Returns:
-        Account -- the local account with this token or a shell account containing
-        just the token and url if no local account is found
-    """
-    accounts = get_local_accounts()
-    if not accounts:
-        return Account.from_token(token, server_url)
-
-    acct = next((acc for acc in accounts if acc.token == token), None)
-    if acct:
-        return acct
-
-    if server_url:
-        url = server_url.lower()
-        acct = next(
-            (acc for acc in accounts if url in acc.serverInfo.url.lower()), None
-        )
-        if acct:
-            return acct
-
-    return Account.from_token(token, server_url)
-
-
-class StreamWrapper:
-    def __init__(self, url: str = None) -> None:
-        raise SpeckleException(
-            message=(
-                "The StreamWrapper has moved as of v2.6.0! Please import from"
-                " specklepy.api.wrapper"
-            ),
-            exception=DeprecationWarning(),
-        )
+
