+++ src/specklepy/api/resources/commit.py
@@ -6,10 +6,10 @@ from specklepy.api.models import Commit
 from specklepy.api.resource import ResourceBase
 from specklepy.logging import metrics
 
-NAME = "commit"
+from specklepy.core.api.resources.commit import NAME, Resource as Core_Resource
 
 
-class Resource(ResourceBase):
+class Resource(Core_Resource):
     """API Access class for commits"""
 
     def __init__(self, account, basepath, client) -> None:
@@ -32,32 +32,9 @@ class Resource(ResourceBase):
         Returns:
             Commit -- the retrieved commit object
         """
-        query = gql(
-            """
-            query Commit($stream_id: String!, $commit_id: String!) {
-                stream(id: $stream_id) {
-                    commit(id: $commit_id) {
-                        id
-                        message
-                        referencedObject
-                        authorId
-                        authorName
-                        authorAvatar
-                        branchName
-                        createdAt
-                        sourceApplication
-                        totalChildrenCount
-                        parents
-                    }
-                }
-            }
-            """
-        )
-        params = {"stream_id": stream_id, "commit_id": commit_id}
+        metrics.track(metrics.COMMIT, self.account, {"name": "get"})
 
-        return self.make_request(
-            query=query, params=params, return_type=["stream", "commit"]
-        )
+        return super().get(stream_id, commit_id)
 
     def list(self, stream_id: str, limit: int = 10) -> List[Commit]:
         """
@@ -70,36 +47,9 @@ class Resource(ResourceBase):
         Returns:
             List[Commit] -- a list of the most recent commit objects
         """
-        metrics.track(metrics.COMMIT, self.account, {"name": "get"})
-        query = gql(
-            """
-            query Commits($stream_id: String!, $limit: Int!) {
-                stream(id: $stream_id) {
-                    commits(limit: $limit) {
-                        items {
-                            id
-                            message
-                            referencedObject
-                            authorName
-                            authorId
-                            authorName
-                            authorAvatar
-                            branchName
-                            createdAt
-                            sourceApplication
-                            totalChildrenCount
-                            parents
-                        }
-                    }
-                }
-            }
-            """
-        )
-        params = {"stream_id": stream_id, "limit": limit}
+        metrics.track(metrics.COMMIT, self.account, {"name": "list"})
 
-        return self.make_request(
-            query=query, params=params, return_type=["stream", "commits", "items"]
-        )
+        return super().list(stream_id, limit)
 
     def create(
         self,
@@ -129,27 +79,8 @@ class Resource(ResourceBase):
             str -- the id of the created commit
         """
         metrics.track(metrics.COMMIT, self.account, {"name": "create"})
-        query = gql(
-            """
-            mutation CommitCreate ($commit: CommitCreateInput!)
-                { commitCreate(commit: $commit)}
-            """
-        )
-        params = {
-            "commit": {
-                "streamId": stream_id,
-                "branchName": branch_name,
-                "objectId": object_id,
-                "message": message,
-                "sourceApplication": source_application,
-            }
-        }
-        if parents:
-            params["commit"]["parents"] = parents
 
-        return self.make_request(
-            query=query, params=params, return_type="commitCreate", parse_response=False
-        )
+        return super().create(stream_id, object_id, branch_name, message, source_application, parents)
 
     def update(self, stream_id: str, commit_id: str, message: str) -> bool:
         """
@@ -165,19 +96,8 @@ class Resource(ResourceBase):
             bool -- True if the operation succeeded
         """
         metrics.track(metrics.COMMIT, self.account, {"name": "update"})
-        query = gql(
-            """
-            mutation CommitUpdate($commit: CommitUpdateInput!)
-                { commitUpdate(commit: $commit)}
-            """
-        )
-        params = {
-            "commit": {"streamId": stream_id, "id": commit_id, "message": message}
-        }
 
-        return self.make_request(
-            query=query, params=params, return_type="commitUpdate", parse_response=False
-        )
+        return super().update(stream_id, commit_id, message)
 
     def delete(self, stream_id: str, commit_id: str) -> bool:
         """
@@ -192,17 +112,8 @@ class Resource(ResourceBase):
             bool -- True if the operation succeeded
         """
         metrics.track(metrics.COMMIT, self.account, {"name": "delete"})
-        query = gql(
-            """
-            mutation CommitDelete($commit: CommitDeleteInput!)
-                { commitDelete(commit: $commit)}
-            """
-        )
-        params = {"commit": {"streamId": stream_id, "id": commit_id}}
 
-        return self.make_request(
-            query=query, params=params, return_type="commitDelete", parse_response=False
-        )
+        return super().delete(stream_id, commit_id)
 
     def received(
         self,
@@ -215,29 +126,4 @@ class Resource(ResourceBase):
         Mark a commit object a received by the source application.
         """
         metrics.track(metrics.COMMIT, self.account, {"name": "received"})
-        query = gql(
-            """
-            mutation CommitReceive($receivedInput:CommitReceivedInput!){
-                commitReceive(input:$receivedInput)
-            }
-            """
-        )
-        params = {
-            "receivedInput": {
-                "sourceApplication": source_application,
-                "streamId": stream_id,
-                "commitId": commit_id,
-                "message": "message",
-            }
-        }
-
-        try:
-            return self.make_request(
-                query=query,
-                params=params,
-                return_type="commitReceive",
-                parse_response=False,
-            )
-        except Exception as ex:
-            print(ex.with_traceback)
-            return False
+        return super().received(stream_id, commit_id, source_application, message)
