+++ src/specklepy/core/api/operations.py
@@ -1,16 +1,11 @@
 from typing import List, Optional
 
-from specklepy.logging import metrics
+#from specklepy.logging import metrics
 from specklepy.logging.exceptions import SpeckleException
 from specklepy.objects.base import Base
 from specklepy.serialization.base_object_serializer import BaseObjectSerializer
 from specklepy.transports.abstract_transport import AbstractTransport
 from specklepy.transports.sqlite import SQLiteTransport
-
-from specklepy.core.api.operations import (send as core_send,
-                                           receive as core_receive,
-                                           serialize as core_serialize,
-                                           deserialize as core_deserialize)
 
 
 def send(
@@ -29,12 +24,27 @@ def send(
     Returns:
         str -- the object id of the sent object
     """
-    obj_hash = core_send(base, transports, use_default_cache)
+
+    if not transports and not use_default_cache:
+        raise SpeckleException(
+            message=(
+                "You need to provide at least one transport: cannot send with an empty"
+                " transport list and no default cache"
+            )
+        )
+
+    if isinstance(transports, AbstractTransport):
+        transports = [transports]
 
     if transports is None:
-        metrics.track(metrics.SEND)
-    else:
-        metrics.track(metrics.SEND, getattr(transports[0], "account", None))
+        transports = []
+
+    if use_default_cache:
+        transports.insert(0, SQLiteTransport())
+
+    serializer = BaseObjectSerializer(write_transports=transports)
+
+    obj_hash, _ = serializer.write_json(base=base)
 
     return obj_hash
 
@@ -55,16 +65,29 @@ def receive(
     Returns:
         Base -- the base object
     """
-    metrics.track(metrics.RECEIVE, getattr(remote_transport, "account", None))
-    return core_receive(obj_id, remote_transport, local_transport)
+    if not local_transport:
+        local_transport = SQLiteTransport()
 
+    serializer = BaseObjectSerializer(read_transport=local_transport)
 
-def _untracked_receive(
-    obj_id: str,
-    remote_transport: Optional[AbstractTransport] = None,
-    local_transport: Optional[AbstractTransport] = None,
-) -> Base:
-    return receive(obj_id, remote_transport, local_transport)
+    # try local transport first. if the parent is there, we assume all the children are there and continue with deserialization using the local transport
+    obj_string = local_transport.get_object(obj_id)
+    if obj_string:
+        return serializer.read_json(obj_string=obj_string)
+
+    if not remote_transport:
+        raise SpeckleException(
+            message=(
+                "Could not find the specified object using the local transport, and you"
+                " didn't provide a fallback remote from which to pull it."
+            )
+        )
+
+    obj_string = remote_transport.copy_object_and_children(
+        id=obj_id, target_transport=local_transport
+    )
+
+    return serializer.read_json(obj_string=obj_string)
 
 
 def serialize(base: Base, write_transports: List[AbstractTransport] = []) -> str:
@@ -81,9 +104,10 @@ def serialize(base: Base, write_transports: List[AbstractTransport] = []) -> str
     Returns:
         str -- the serialized object
     """
-    metrics.track(metrics.SERIALIZE)
+    serializer = BaseObjectSerializer(write_transports=write_transports)
 
-    return core_serialize(base, write_transports)
+    return serializer.write_json(base)[1]
+
 
 def deserialize(
     obj_string: str, read_transport: Optional[AbstractTransport] = None
@@ -104,9 +128,12 @@ def deserialize(
     Returns:
         Base -- the deserialized object
     """
-    metrics.track(metrics.DESERIALIZE)
+    if not read_transport:
+        read_transport = SQLiteTransport()
 
-    return core_deserialize(obj_string, read_transport)
+    serializer = BaseObjectSerializer(read_transport=read_transport)
+
+    return serializer.read_json(obj_string=obj_string)
 
 
 __all__ = ["receive", "send", "serialize", "deserialize"]
