+++ src/specklepy/transports/server/server.py
@@ -5,7 +5,7 @@ from warnings import warn
 import requests
 
 from specklepy.api.client import SpeckleClient
-from specklepy.api.credentials import Account, get_account_from_token
+from specklepy.core.api.credentials import Account, get_account_from_token
 from specklepy.logging.exceptions import SpeckleException, SpeckleWarning
 from specklepy.transports.abstract_transport import AbstractTransport
